+++ entities/exercise_choice.go
@@ -3,8 +3,8 @@ package entities
 import "time"
 
 type ExerciseChoice struct {
-	ID         int       `gorm:"column:id"`
-	ExerciseID int       `gorm:"column:exercise_id"`
+	ID         uint64    `gorm:"column:id"`
+	ExerciseID uint64    `gorm:"column:exercise_id"`
 	ChoiceText string    `gorm:"column:choice_text"`
 	IsCorrect  bool      `gorm:"column:is_correct"`
 	CreatedAt  time.Time `gorm:"column:created_at"`
